# [`Fix`] References in `StorageItem`