@Jim8y 2025-02-25
This pr looks fixed the claimed issue, and originally it can not ensure the immutbality of the content passed in, new impl ensures that every storage item has its own copy of the content. For security consideration, its OK, but will sacrifice the performance i think. Maybe we can keep the original one as well for hotspot.