@cschuchardt88 2025-02-25
> This pr looks fixed the claimed issue, and originally it can not ensure the immutbality of the content passed in, new impl ensures that every storage item has its own copy of the content. For security consideration, its OK, but will sacrifice the performance i think. Maybe we can keep the original one as well for hotspot.

It would seem like the copy is slow. But in fact `Memory` and `Span` is like using `byte*`, there just pointers to an address and `GC` doesn't move them around, Like it does for objects like `byte[]`,  `GC`  will try to optimize the object by pushing it off and on stack, unlike `Memory` and `Span`, where they're fixed. This is 50% performance increase if I remember correctly. Making these copies very fast. If it becomes a problem then we can create a memory pool, with `IMemoryOwner` like so:

```csharp
internal sealed class MemoryPoolBlock : IMemoryOwner<byte>
    {
        public PinnedBlockMemoryPool Pool { get; }

        internal MemoryPoolBlock(
            PinnedBlockMemoryPool pool,
            int length)
        {
            Pool = pool;

            var pinnedArray = GC.AllocateUninitializedArray<byte>(length, pinned: true);

            Memory = MemoryMarshal.CreateFromPinnedArray(pinnedArray, 0, pinnedArray.Length);
        }

        #region IMemoryOwner

        public Memory<byte> Memory { get; }

        void IDisposable.Dispose()
        {
            Pool.Return(this);
        }

        #endregion
    }
```