@cschuchardt88 2025-02-26
These benchmarks are with random data. The performance is only nano-seconds. The point of this PR If you recall is to have `StorageItem` be `immutable`. Currently it is not and can be a problem. If you don't want it to be `immutable` than you should be passing as `ref` object. This is not the case and not the intent of the original developer, seeing how they used `ref` in place where they wanted the data to be changed.

```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3194)
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK 9.0.200
  [Host]     : .NET 9.0.2 (9.0.225.6610), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.2 (9.0.225.6610), X64 RyuJIT AVX2
```
## After
| Method      | Mean      | Error    | StdDev   |
|------------ |----------:|---------:|---------:|
| CreateNew   |  56.94 ns | 0.926 ns | 0.866 ns |
| FromReplica |  83.44 ns | 1.112 ns | 0.986 ns |
| SetValue    | 113.32 ns | 1.578 ns | 1.232 ns |
| Deserialize | 186.53 ns | 2.404 ns | 2.008 ns |

## Before
| Method      | Mean        | Error     | StdDev    |
|------------ |------------:|----------:|----------:|
| CreateNew   |   0.0096 ns | 0.0077 ns | 0.0060 ns |
| FromReplica |  17.8435 ns | 0.1501 ns | 0.1253 ns |
| SetValue    |   0.2327 ns | 0.0049 ns | 0.0044 ns |
| Deserialize |   0.8554 ns | 0.0082 ns | 0.0073 ns |