@cschuchardt88 2025-02-27
>Why are the before data so fast.

CPUs can perform certain operations more efficiently through techniques like pipelining, superscalar execution, and parallel processing, out of order execution, cache memory  (L1, L2, L3), branch prediction and speculative execution.