@cschuchardt88 2025-02-27
> Pipeling, OoO and other mechanisms can make it possible for many instructions to be executed less than the total time for each individual to be executed(i.e. execution time of 1000 instructions less then 1000 cycles). But cannot less than one cycle.


That is true, that you can't be faster than 1 cycle. But that's how it can look faster than 1 cycle. Just mis-calculation.