@nan01ab 2025-02-27
> > > > Why are the before data so fast.
> > > 
> > > 
> > > CPUs can perform certain operations more efficiently through techniques like pipelining, superscalar execution, and parallel processing, out of order execution, cache memory (L1, L2, L3), branch prediction and speculative execution.
> > 
> > 
> > Even if many mechanisms make execution faster, An operation cannot faster than A single CPU cycle.
> > Pipeling, OoO and other mechanisms can make it possible for many instructions to be executed less than the total time for each individual to be executed(i.e. execution time of 1000 instructions less then 1000 cycles). But cannot less than one cycle.
> 
> That situation actually happens a lot while i use NetBenchmark...
 Because the operation has no actual effect on something, it may be optimized to nothing.