@Jim8y 2025-02-28
> A summary from regular core dev meeting: we should try to synchronize Mainnet node from dump using this PR and check how this PR affects the overall synchronization time comparing to master.

to be honest, if this is how we deals with it, then all pr will request something like this.