@shargon 2025-02-28
> > A summary from regular core dev meeting: we should try to synchronize Mainnet node from dump using this PR and check how this PR affects the overall synchronization time comparing to master.
> 
> to be honest, if this is how we deal with it, then all pr will request something like this, i dont see why this pr is so special.

For me it's enough with real scenario benchmark, i see more copies than before, some of them I don't understand why, if it's already read only memory