@cschuchardt88 2025-02-28
> What's the problem that this pr try to solve?

Tell you monday.

```csharp
byte[] bytes = [0xff];
var mem = new ReadOnlyMemory<byte>(bytes);

bytes[0] = 0xdd; // this changes the value in mem variable.

Assert.IsTrue(mem.ToArray()[0] == 0xdd);
```