@shargon 2025-03-01
> > What's the problem that this pr try to solve?
> 
> Tell you monday.
> 
> This is bad for `StorageKey` and `StorageItem`
> 
> ```cs
> [TestMethod]
> public void Test()
> {
>     byte[] bytes = [0xff];
>     var mem = new ReadOnlyMemory<byte>(bytes);
> 
>     bytes[0] = 0xdd; // this changes the value in mem variable.
> 
>     Assert.IsTrue(mem.ToArray()[0] == 0xdd);
> }
> ```
> 
> ![image](https://private-user-images.githubusercontent.com/8141309/418034204-004d5e0b-0fca-4bd6-aeed-cffa4db94dc7.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDA4MDExNTMsIm5iZiI6MTc0MDgwMDg1MywicGF0aCI6Ii84MTQxMzA5LzQxODAzNDIwNC0wMDRkNWUwYi0wZmNhLTRiZDYtYWVlZC1jZmZhNGRiOTRkYzcucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDMwMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAzMDFUMDM0NzMzWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9NjUyODQ3NjY4YTQwNTI2MGEyMTU5YzllMGE1ZWJjNjJmOGQyN2NjMTdjY2NlM2JlMDU2NmIwOWNjZWJlZGFkNyZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.lK7MuB5evwDeKsONjTf2jvVNE70mNPeKsaKqCOdiKts)

I understand you, but it was noticed in the constructor's comment, make a copy will make it too much slower, maybe we should create different constructors but i didn't see any risk with the current code, also the deserialize is not affected by this, I'm sure we can find a better way than clone the current data.