@shargon 2025-03-01
> Somewhere this is going to impact performance anyways, if we don't do it. We are only talking about 20-100 nanoseconds. We need this change for added security. Someone can corrupt their `store` (_`DataCache`_).

Is the same with the script of the transaction and any byte array that we use.