@cschuchardt88 2025-03-01
@AnnaShaleva @shargon @nan01ab @Jim8y 

### Benchmark Results (offline sync)
100,000 blocks/94secs - Before
100,000 blocks/92secs - After with #3783 #3782